@@ -1,7 +1,7 @@
 {
   "name": "boilerplate_frontend_react",
   "private": false,
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "Boilerplate built with react, vite, sass, cypress, cucumber, eslint & stylelint",
   "web": "dist/index.html",
   "targets": {
@@ -38,38 +38,38 @@
     "copy": "cp src/assets/_redirects dist"
   },
   "dependencies": {
-    "@vitejs/plugin-react": "^1.2.0",
-    "i18next": "^21.6.14",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-i18next": "^11.15.6",
-    "react-router-dom": "^6.2.2",
-    "rollup-plugin-visualizer": "^5.6.0",
-    "vite": "^2.8.0",
-    "vite-plugin-environment": "^1.1.0"
+    "@vitejs/plugin-react": "^4.0.4",
+    "i18next": "^23.4.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-i18next": "^13.2.0",
+    "react-router-dom": "^6.15.0",
+    "rollup-plugin-visualizer": "^5.9.2",
+    "vite": "^4.4.9",
+    "vite-plugin-environment": "^1.1.3"
   },
   "devDependencies": {
-    "cypress": "^9.5.1",
+    "@cypress/xpath": "^2.0.3",
+    "cypress": "^12.17.4",
     "cypress-cucumber-preprocessor": "^4.3.1",
-    "cypress-xpath": "^1.6.2",
-    "eslint": "^8.10.0",
+    "eslint": "^8.48.0",
     "eslint-config-cypress": "^0.28.0",
-    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-prettier": "^9.0.0",
     "eslint-config-react": "^1.1.7",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.29.3",
-    "http-server": "^14.1.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-react": "^7.33.2",
+    "http-server": "^14.1.1",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.5.1",
-    "rimraf": "^3.0.2",
-    "sass": "^1.49.9",
-    "stylelint": "^14.5.3",
-    "stylelint-config-sass-guidelines": "^9.0.1",
-    "stylelint-config-standard": "^25.0.0",
-    "stylelint-config-standard-scss": "^3.0.0",
-    "stylelint-scss": "^4.1.0"
+    "prettier": "^3.0.2",
+    "rimraf": "^5.0.1",
+    "sass": "^1.66.1",
+    "stylelint": "^15.10.3",
+    "stylelint-config-sass-guidelines": "^10.0.0",
+    "stylelint-config-standard": "^34.0.0",
+    "stylelint-config-standard-scss": "^10.0.0",
+    "stylelint-scss": "^5.1.0"
   },
   "cypress-cucumber-preprocessor": {
-    "nonGlobalStepDefinitions": true
+    "nonGlobalStepDefinitions": true 
   }
 }
