@@ -1,5 +1,5 @@
 import { Given } from "cypress-cucumber-preprocessor/steps";
 
-Given('I open {string} page', url => {
-    cy.visit(url)
-})
+Given("I open {string} page", url => {
+  cy.visit(url);
+});
